Blank maze creator function implemented
first commit of the project, all that this commit is really adding is a function that will create a nested list with all cells marked as unvisited. This commit also happens to include functionality to print out that array in a more visual manner, but that is a temporary solution.