Adjacent Cells Function
The function added in this commit accepts the height and width parameters along with the coordinate points of a cell in list form as input. In return the function calculates all the adjacent cells that are not on the border of the maze and returns those cells to the caller in list form.