Added non-functional GUI components for stage 1
The Gui is now equipped with a crappy title, labelled entry fields for width and height parameters, as well as a button that is meant to execute the building process. As of yet all of the above mentioned GUI interaction points are not functional. But non-functional is such an ugly way of putting it, instead let's say that blank canvases have been lovingly primed and prepared for a dash of life to be breathed into them.