Entry fields now respond to user input properly
When the build maze button is pressed now the inputted values taken in by the entry fields are used as actual dimensions for building the maze the user has actually requested. There has been a default maze set when the program opens, a grey 15 by 15 grid meant to represent an empty maze. In addition to this default graphical output there has also been an addition of two default values to the entry fields that correspond to the dimensions of the default maze. They are functional in the capacity that if you press the button straight out of the gate they generate a maze with the proper dimensions